Moved adapter tests to own test package. Updated PaperAdapterTests. Fixed wrong parameter name in IAdapter.
