+++ ApplicationLogicTests/ApplicationLogicTests.csproj
@@ -131,11 +131,11 @@
     <Compile Include="Mapping\AutoMapperConfiguratorTest.cs" />
     <Compile Include="PaperManagement\Savers\BibtexTagSaverTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UserManagement\TeamAdapterTests.cs" />
+    <Compile Include="StorageAdapter\TeamAdapterTests.cs" />
     <Compile Include="UserManagement\TeamValidatorTests.cs" />
     <Compile Include="UserManagement\TeamHandlerTests.cs" />
     <Compile Include="UserManagement\UserHandlerTest.cs" />
-    <Compile Include="UserManagement\UserAdapterTests.cs" />
+    <Compile Include="StorageAdapter\UserAdapterTests.cs" />
     <Compile Include="UserManagement\UserTest.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -160,7 +160,7 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <Compile Include="Resources\StorageFasade\PaperFacadeTests.cs" />
+    <Compile Include="StorageAdapter\PaperAdapterTests.cs" />
     <Compile Include="StudyManagement\CriteriaFactoryTest.cs" />
   </ItemGroup>
   <ItemGroup>
